Use a web-friendly URL in post-deploy message
GitHub appears to try and parse the "@" username included in the message, and hides everything past the "@" symbol entirely (potentially because my expo account username and github username is the same?). Expo does provide a hosted web link that can be linked to instead - which is probably more helpful then the `expo://` url anyways (since it's tappable and includes instructions for downloading expo, etc).